tg-10424 [BE] Unidad/Dosis y Dosis/Dia deben ser campos de punto flotante

+++ back-end/clinic-history/src/main/java/ar/lamansys/sgh/clinichistory/domain/ips/DosageBo.java
@@ -34,11 +34,11 @@ public class DosageBo {
 
 	private QuantityBo quantity;
 
-	private Integer dosesByUnit;
+	private Double dosesByUnit;
 
-	private Integer dosesByDay;
+	private Double dosesByDay;
 
-	public DosageBo(Double duration, Integer dosesByUnit, Integer dosesByDay) {
+	public DosageBo(Double duration, Double dosesByUnit, Double dosesByDay) {
 		this.duration = duration;
 		this.dosesByUnit = dosesByUnit;
 		this.dosesByDay = dosesByDay;
