+++ back-end/clinic-history/src/main/java/ar/lamansys/sgh/clinichistory/infrastructure/output/repository/ips/entity/Dosage.java
@@ -60,10 +60,10 @@ public class Dosage {
 	private String event;
 
 	@Column(name = "doses_by_day")
-	private Integer dosesByDay;
+	private Double dosesByDay;
 
 	@Column(name = "doses_by_unit")
-	private Integer dosesByUnit;
+	private Double dosesByUnit;
 
 
 	public Dosage(Double duration, String durationUnit,
